refactor: Move user code to a specific package
The current code structure is tied together despite of the packages or
actors logical separation, which can become a mess when the code grows
up. In order to solve this problem, this patch moves the user's related
code to a specific package/folder.

This change follows the golang project structure recommendation, adopted
by a lot of open-source projects.
https://github.com/golang-standards/project-layout/tree/master/pkg
