+++ pkg/user/interactors/create_user.go
@@ -1,19 +1,19 @@
 package interactors
 
 import (
-	"github.com/CESARBR/knot-babeltower/pkg/entities"
 	"github.com/CESARBR/knot-babeltower/pkg/logging"
-	"github.com/CESARBR/knot-babeltower/pkg/network"
+	"github.com/CESARBR/knot-babeltower/pkg/user/delivery/http"
+	"github.com/CESARBR/knot-babeltower/pkg/user/entities"
 )
 
 // CreateUser to interact to user
 type CreateUser struct {
 	logger    logging.Logger
-	userProxy network.UserProxy
+	userProxy http.UserProxy
 }
 
 // NewCreateUser contructs the interactor
-func NewCreateUser(logger logging.Logger, userProxy network.UserProxy) *CreateUser {
+func NewCreateUser(logger logging.Logger, userProxy http.UserProxy) *CreateUser {
 	return &CreateUser{logger, userProxy}
 }
